Fixed comments on PR
