+++ app/[lang]/dictionaries.ts
@@ -1,3 +1,9 @@
+interface LangDictionary {
+  en: () => Promise<Dict>;
+  fr: () => Promise<Dict>;
+  kw: () => Promise<Dict>;
+}
+
 const dictionaries: LangDictionary = {
   en: () => import('./dictionaries/en.json').then((module) => module.default),
   fr: () => import('./dictionaries/fr.json').then((module) => module.default),
@@ -5,11 +11,7 @@ const dictionaries: LangDictionary = {
 };
 
 export const getDictionary = async (lang: Lang) => {
-  const dictionaryFunction = dictionaries[lang];
-  if (typeof dictionaryFunction === 'function') {
-    return await dictionaryFunction();
-  }
-  return {}; // Return an empty object or handle the error as needed.
+  return dictionaries[lang]();
 };
 
 export const getFromDictionary = (
