+++ app/[lang]/dictionaries/en.json
@@ -1,4 +1,9 @@
 {
+  "title": "Floodsafe",
   "createAccount": "Create account",
-  "logIn": "Log in"
+  "logIn": "Log in",
+  "additionalInfo": "Some additional explanatory text here, this is a placeholder.",
+  "firstOnboardingPoint": "Receive flood warnings",
+  "secondOnboardingPoint": "Some other point",
+  "thirdOnboardingPoint": "And one final nice thing"
 }
