+++ app/[lang]/dictionaries/fr.json
@@ -1,4 +1,9 @@
 {
+  "title": "Floodsafe",
   "createAccount": "Créer un compte",
-  "logIn": "Se connecter"
+  "logIn": "Se connecter",
+  "additionalInfo": "Some additional explanatory text here, this is a placeholder.",
+  "firstOnboardingPoint": "Receive flood warnings",
+  "secondOnboardingPoint": "Some other point",
+  "thirdOnboardingPoint": "And one final nice thing"
 }
