+++ app/[lang]/intro-screen/page.tsx
@@ -2,8 +2,8 @@
 
 import { Box } from '@mui/material';
 import bg from '@/public/assets/images/start-screen-image.png';
-import Title from '@/app/[lang]/components/Title';
-import LanguageModal from '@/app/[lang]/components/LanguageModal';
+import Title from '@/app/components/Title';
+import LanguageModal from '@/app/components/LanguageModal';
 import { useState } from 'react';
 import { locales } from '@/middleware';
 import { useRouter } from 'next/navigation';
@@ -16,9 +16,10 @@ interface IntroScreenProps {
     value: any,
     options?: CookieSetOptions | undefined
   ) => void;
+  dict: Dict;
 }
 
-const IntroScreen = ({ cookies, setCookie }: IntroScreenProps) => {
+const IntroScreen = ({ cookies, setCookie, dict }: IntroScreenProps) => {
   const [isModalOpen, setIsModalOpen] = useState(!cookies.language);
   const router = useRouter();
 
@@ -28,6 +29,7 @@ const IntroScreen = ({ cookies, setCookie }: IntroScreenProps) => {
       router.replace('/' + localeString);
     }
   };
+
   return (
     <Box style={{ height: '100%', width: '100%' }}>
       <Box
@@ -41,7 +43,7 @@ const IntroScreen = ({ cookies, setCookie }: IntroScreenProps) => {
           height: '100%',
         }}
       >
-        <Title />
+        <Title dict={dict} />
         <LanguageModal
           open={isModalOpen}
           onClose={() => setIsModalOpen(false)}
