+++ app/[lang]/types.d.ts
@@ -1,13 +1,3 @@
-interface PageProps {
-  lang: string;
-}
-
-interface LangDictionary {
-  en: () => Promise<Dict>;
-  fr: () => Promise<Dict>;
-  kw: () => Promise<Dict>;
-}
-
 type Lang = 'en' | 'fr' | 'kw';
 
 type Dict = {
