+++ app/components/Title.tsx
@@ -1,7 +1,12 @@
 import { Box, Typography } from '@mui/material';
 import Logo from '@/public/assets/icons/Logo';
+import { t } from '@/app/[lang]/dictionaries';
 
-const Title = () => {
+interface TitleProps {
+  dict: Dict;
+}
+
+const Title = ({ dict }: TitleProps) => {
   return (
     <Box
       style={{
@@ -20,7 +25,7 @@ const Title = () => {
         variant={'h4'}
         style={{ marginLeft: '10px' }}
       >
-        Floodsafe
+        {t('title', dict)}
       </Typography>
     </Box>
   );
