+++ public/assets/icons/Logo.tsx
@@ -1,19 +1,17 @@
 const Logo = () => {
   return (
-    <>
-      <svg
-        width="41"
-        height="23"
-        viewBox="0 0 41 23"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          fill={'black'}
-          d="M7.50665 12.4792C9.84364 12.4792 10.2953 14.4376 14.0463 14.4376C17.8758 14.4376 18.1507 12.4792 20.5859 12.4792C22.9229 12.4792 23.3746 14.4376 27.1256 14.4376C30.9551 14.4376 31.23 12.4792 33.6652 12.4792C36.0022 12.4792 36.4146 14.3984 40.1656 14.4376V10.5209C37.8286 10.5209 37.3769 8.56258 33.6259 8.56258C29.7964 8.56258 29.5215 10.5209 27.0863 10.5209C24.7493 10.5209 24.2976 8.56258 20.5466 8.56258C16.7171 8.56258 16.4422 10.5209 14.007 10.5209C11.67 10.5209 11.2183 8.56258 7.46737 8.56258C3.63785 8.56258 3.36292 10.5209 0.927734 10.5209V14.4376C4.65906 14.4376 5.1893 12.4792 7.50665 12.4792ZM33.6652 16.3959C29.8357 16.3959 29.5607 18.3542 27.1256 18.3542C24.7886 18.3542 24.3369 16.3959 20.5859 16.3959C16.7564 16.3959 16.4618 18.3542 14.0266 18.3542C11.5915 18.3542 11.3165 16.3959 7.48701 16.3959C3.65749 16.3959 3.36292 18.3542 0.927734 18.3542V22.2709C4.75725 22.2709 5.07147 20.3126 7.48701 20.3126C9.92219 20.3126 10.1971 22.2709 14.0266 22.2709C17.8562 22.2709 18.1507 20.3126 20.5859 20.3126C22.9229 20.3126 23.3746 22.2709 27.1256 22.2709C30.9354 22.2709 31.23 20.3126 33.6652 20.3126C36.0022 20.3126 36.4539 22.2709 40.2048 22.2709V18.3542C37.7697 18.3542 37.4947 16.3959 33.6652 16.3959ZM7.50665 4.64591C9.84364 4.64591 10.2953 6.60425 14.0463 6.60425C17.8758 6.60425 18.1507 4.64591 20.5859 4.64591C22.9229 4.64591 23.3746 6.60425 27.1256 6.60425C30.9551 6.60425 31.23 4.64591 33.6652 4.64591C36.0022 4.64591 36.4146 6.56508 40.1656 6.60425V2.68758C37.8286 2.68758 37.3769 0.729248 33.6259 0.729248C29.7964 0.729248 29.5215 2.68758 27.0863 2.68758C24.7493 2.68758 24.2976 0.729248 20.5466 0.729248C16.7171 0.729248 16.4422 2.68758 14.007 2.68758C11.67 2.68758 11.2183 0.729248 7.46737 0.729248C3.63785 0.729248 3.36292 2.68758 0.927734 2.68758V6.60425C4.65906 6.60425 5.1893 4.64591 7.50665 4.64591Z"
-        />
-      </svg>
-    </>
+    <svg
+      width="41"
+      height="23"
+      viewBox="0 0 41 23"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fill={'black'}
+        d="M7.50665 12.4792C9.84364 12.4792 10.2953 14.4376 14.0463 14.4376C17.8758 14.4376 18.1507 12.4792 20.5859 12.4792C22.9229 12.4792 23.3746 14.4376 27.1256 14.4376C30.9551 14.4376 31.23 12.4792 33.6652 12.4792C36.0022 12.4792 36.4146 14.3984 40.1656 14.4376V10.5209C37.8286 10.5209 37.3769 8.56258 33.6259 8.56258C29.7964 8.56258 29.5215 10.5209 27.0863 10.5209C24.7493 10.5209 24.2976 8.56258 20.5466 8.56258C16.7171 8.56258 16.4422 10.5209 14.007 10.5209C11.67 10.5209 11.2183 8.56258 7.46737 8.56258C3.63785 8.56258 3.36292 10.5209 0.927734 10.5209V14.4376C4.65906 14.4376 5.1893 12.4792 7.50665 12.4792ZM33.6652 16.3959C29.8357 16.3959 29.5607 18.3542 27.1256 18.3542C24.7886 18.3542 24.3369 16.3959 20.5859 16.3959C16.7564 16.3959 16.4618 18.3542 14.0266 18.3542C11.5915 18.3542 11.3165 16.3959 7.48701 16.3959C3.65749 16.3959 3.36292 18.3542 0.927734 18.3542V22.2709C4.75725 22.2709 5.07147 20.3126 7.48701 20.3126C9.92219 20.3126 10.1971 22.2709 14.0266 22.2709C17.8562 22.2709 18.1507 20.3126 20.5859 20.3126C22.9229 20.3126 23.3746 22.2709 27.1256 22.2709C30.9354 22.2709 31.23 20.3126 33.6652 20.3126C36.0022 20.3126 36.4539 22.2709 40.2048 22.2709V18.3542C37.7697 18.3542 37.4947 16.3959 33.6652 16.3959ZM7.50665 4.64591C9.84364 4.64591 10.2953 6.60425 14.0463 6.60425C17.8758 6.60425 18.1507 4.64591 20.5859 4.64591C22.9229 4.64591 23.3746 6.60425 27.1256 6.60425C30.9551 6.60425 31.23 4.64591 33.6652 4.64591C36.0022 4.64591 36.4146 6.56508 40.1656 6.60425V2.68758C37.8286 2.68758 37.3769 0.729248 33.6259 0.729248C29.7964 0.729248 29.5215 2.68758 27.0863 2.68758C24.7493 2.68758 24.2976 0.729248 20.5466 0.729248C16.7171 0.729248 16.4422 2.68758 14.007 2.68758C11.67 2.68758 11.2183 0.729248 7.46737 0.729248C3.63785 0.729248 3.36292 2.68758 0.927734 2.68758V6.60425C4.65906 6.60425 5.1893 4.64591 7.50665 4.64591Z"
+      />
+    </svg>
   );
 };
 export default Logo;
